requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
alembic==1.13.1
scipy==1.11.4
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
